// frontend/src/entities/structure/types/mutations.ts
export enum MutationTypes {
    SET_BLOCK_AREA = "SET_BLOCK_AREA",
    SET_ELEMENT = "SET_ELEMENT",
    PUSH_CHILD_TO_BLOCK_AREA = "PUSH_CHILD_TO_BLOCK_AREA",
    PUSH_WINDOW_TO_BLOCK_AREA = "PUSH_WINDOW_TO_BLOCK_AREA",
    SPLICE_CHILD_TO_BLOCK_AREA = "SPLICE_CHILD_TO_BLOCK_AREA",
    SET_CHILDREN_TO_BLOCK_AREA = "SET_CHILDREN_TO_BLOCK_AREA",
    DELETE_ELEMENT = "DELETE_ELEMENT",
    DELETE_CHILD = "DELETE_CHILD",
    DELETE_WINDOW = "DELETE_WINDOW",
}